move the model and helper functions to separate files
